moved enemy related classes to own folder
