+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/Enemy/Enemy.java
@@ -1,5 +1,7 @@
-package com.github.migi_1.Context.model.entity;
+package com.github.migi_1.Context.model.Enemy;
 
+import com.github.migi_1.Context.model.entity.Entity;
+import com.github.migi_1.Context.model.entity.IKillable;
 import com.github.migi_1.Context.utility.ProjectAssetManager;
 import com.jme3.math.Vector3f;
 import com.jme3.scene.Spatial;
