+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/Enemy/EnemyGenerator.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.Context.model.entity;
+package com.github.migi_1.Context.model.Enemy;
 
 import java.util.LinkedList;
